pyproject.toml
[project]
name = "xhs_public_opinion"
version = "0.1.0"
description = "xhs_public_opinion using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.114.0,<1.0.0",
    "supabase>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "pandas>=2.0.0"
]

[project.scripts]
xhs_public_opinion = "xhs_public_opinion.main:run"
run_crew = "xhs_public_opinion.main:run"
train = "xhs_public_opinion.main:train"
replay = "xhs_public_opinion.main:replay"
test = "xhs_public_opinion.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
